Re-factor common datetime types from metar into datetime module
